+++ CHANGES.md
@@ -30,6 +30,7 @@
    model for comparison with the data.  The new approach should nominally be
    closer to the effects of binning the spectra and then fitting the kinematics.
  - Added method to compute Fisher Information Matrix for axisymmetric fits.
+ - Added new asymmetry map calculations
 
 
 0.1.0